scripts/preprocess_all.py
"""
=============================================================================
Complete Preprocessing Pipeline Script
=============================================================================
This script runs the complete preprocessing pipeline:
1. DICOM to NIfTI conversion
2. RTSTRUCT mask extraction
3. Data normalization
4. Dataset splitting

Usage:
    python preprocess_all.py --config src/config/params.yaml

Author: Medical Imaging Team
Date: November 2025
=============================================================================
"""

import os
import sys
import argparse
import yaml
import logging
from pathlib import Path

# Add src to path
sys.path.append(str(Path(__file__).parent.parent / 'src'))

from preprocessing import (
    convert_dicom_to_nifti,
    extract_mask_from_rtstruct,
    normalize_data,
    split_dataset
)

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def main():
    """Main preprocessing pipeline."""
    parser = argparse.ArgumentParser(description='Run complete preprocessing pipeline')
    parser.add_argument('--config', type=str, default='src/config/params.yaml',
                        help='Path to configuration file')
    parser.add_argument('--step', type=str, default='all',
                        choices=['all', 'convert', 'extract', 'normalize', 'split'],
                        help='Which preprocessing step to run')
    args = parser.parse_args()
    
    # Load configuration
    config = load_config(args.config)
    logger.info(f"Loaded configuration from {args.config}")
    
    # TODO: Implement complete preprocessing pipeline
    logger.info("Preprocessing pipeline not yet implemented")
    logger.info("This will be completed in Phase 2-3")
    
    # Step 1: DICOM to NIfTI conversion
    if args.step in ['all', 'convert']:
        logger.info("Step 1: DICOM to NIfTI conversion")
        # TODO: Call conversion functions
        pass
    
    # Step 2: RTSTRUCT mask extraction
    if args.step in ['all', 'extract']:
        logger.info("Step 2: RTSTRUCT mask extraction")
        # TODO: Call mask extraction functions
        pass
    
    # Step 3: Data normalization
    if args.step in ['all', 'normalize']:
        logger.info("Step 3: Data normalization")
        # TODO: Call normalization functions
        pass
    
    # Step 4: Dataset splitting
    if args.step in ['all', 'split']:
        logger.info("Step 4: Dataset splitting")
        # TODO: Call splitting functions
        pass
    
    logger.info("Preprocessing complete!")


if __name__ == "__main__":
    main()
