src/preprocessing/split_dataset.py
"""
=============================================================================
Dataset Splitting Module
=============================================================================
This module handles train/validation/test splitting.

Functions:
    - load_patient_list: Load list of available patients
    - stratified_split: Split with stratification
    - random_split: Random split
    - save_splits: Save split information
    - load_splits: Load existing split information

Author: Medical Imaging Team
Date: November 2025
=============================================================================
"""

import os
import json
import numpy as np
import pandas as pd
from pathlib import Path
from typing import Tuple, List, Dict, Optional
from sklearn.model_selection import train_test_split
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def load_patient_list(data_dir: str) -> List[str]:
    """
    Load list of available patients from data directory.
    
    Args:
        data_dir: Directory containing patient data
        
    Returns:
        List of patient IDs
    """
    # TODO: Implement patient list loading
    pass


def stratified_split(
    patient_ids: List[str],
    labels: List[int],
    train_ratio: float = 0.70,
    val_ratio: float = 0.15,
    test_ratio: float = 0.15,
    random_seed: int = 42
) -> Dict[str, List[str]]:
    """
    Perform stratified train/val/test split.
    
    Args:
        patient_ids: List of patient identifiers
        labels: Labels for stratification
        train_ratio: Training set ratio
        val_ratio: Validation set ratio
        test_ratio: Test set ratio
        random_seed: Random seed for reproducibility
        
    Returns:
        Dictionary with 'train', 'val', 'test' patient lists
    """
    # TODO: Implement stratified splitting
    pass


def random_split(
    patient_ids: List[str],
    train_ratio: float = 0.70,
    val_ratio: float = 0.15,
    test_ratio: float = 0.15,
    random_seed: int = 42
) -> Dict[str, List[str]]:
    """
    Perform random train/val/test split.
    
    Args:
        patient_ids: List of patient identifiers
        train_ratio: Training set ratio
        val_ratio: Validation set ratio
        test_ratio: Test set ratio
        random_seed: Random seed for reproducibility
        
    Returns:
        Dictionary with 'train', 'val', 'test' patient lists
    """
    # TODO: Implement random splitting
    pass


def save_splits(
    splits: Dict[str, List[str]],
    output_path: str
) -> None:
    """
    Save split information to JSON file.
    
    Args:
        splits: Dictionary with train/val/test splits
        output_path: Output JSON file path
    """
    # TODO: Implement split saving
    pass


def load_splits(split_path: str) -> Dict[str, List[str]]:
    """
    Load existing split information from JSON file.
    
    Args:
        split_path: Path to split JSON file
        
    Returns:
        Dictionary with train/val/test splits
    """
    # TODO: Implement split loading
    pass


def main():
    """
    Main function to perform dataset splitting.
    """
    # TODO: Implement main splitting logic
    pass


if __name__ == "__main__":
    logger.info("Dataset splitting module loaded")
    logger.info("This module will be implemented in Phase 3")
    main()
