requirements.txt
# ============================================================================
# REQUIREMENTS.TXT - NSCLC Segmentation Project
# ============================================================================
# Install: pip install -r requirements.txt
# ============================================================================

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0

# Medical imaging
pydicom>=2.3.0
SimpleITK>=2.2.0
nibabel>=5.1.0
rt-utils>=1.2.7
pydicom-seg>=0.4.1
monai>=1.2.0

# Image processing
scikit-image>=0.21.0
opencv-python>=4.7.0
Pillow>=9.5.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Segmentation models
segmentation-models-pytorch>=0.3.3

# Data handling
h5py>=3.8.0
PyYAML>=6.0

# Progress and utilities
tqdm>=4.65.0
colorama>=0.4.6
rich>=13.3.5

# Jupyter
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Experiment tracking
tensorboard>=2.13.0
tensorboardX>=2.6.0
# wandb>=0.15.0  # Uncomment if using Weights & Biases

# Machine learning utilities
scikit-learn>=1.2.0

# Environment
python-dotenv>=1.0.0

# Testing (optional)
pytest>=7.3.0
pytest-cov>=4.1.0
